git-tfs-id: [http://uk-tfs02:8080/tfs/DefaultCollection]$/Erm.Core/Dev/Features/OVR/BLFlex;C24499

+++ Source/ApplicationServices/2Gis.Erm.BLFlex.Operations.Global/Shared/Formatter/DirhamInEnglishPluralizer.cs
@@ -4,7 +4,7 @@ namespace DoubleGis.Erm.BLFlex.Operations.Global.Shared.Formatter
 {
     public sealed class DirhamInEnglishPluralizer : IWordPluralizer
     {
-        public string GetPluralFor(int numberToRead)
+        public string GetPluralFor(long numberToRead)
         {
             return Math.Abs(numberToRead) == 1 ? "dirham" : "dirhams";
         }
